requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# HuggingFace
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
evaluate>=0.4.0

# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0

# For Russian language models
sentencepiece>=0.1.99
protobuf>=4.24.0

# Optional: for experiment tracking
# wandb>=0.15.0’’